+++ language/src/main/java/i4gl/runtime/values/Char.java
@@ -2,18 +2,15 @@ package i4gl.runtime.values;
 
 import java.util.Arrays;
 
+import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
-import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.InvalidArrayIndexException;
 import com.oracle.truffle.api.interop.TruffleObject;
 import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
 
-import i4gl.I4GLLanguage;
 import i4gl.exceptions.NotImplementedException;
-import i4gl.runtime.context.Context;
-import i4gl.runtime.types.compound.Char1Type;
-import i4gl.runtime.types.compound.CharType;
 
 /**
  * Representation of variables of NChar type. It is a slight wrapper to Java's
@@ -113,35 +110,49 @@ public class Char implements TruffleObject {
     }
 
     @ExportMessage
-    boolean hasLanguage() {
+    boolean hasArrayElements() {
         return true;
     }
 
     @ExportMessage
-    Class<? extends TruffleLanguage<Context>> getLanguage() {
-        return I4GLLanguage.class;
+    final boolean isArrayElementInsertable(long index) {
+        return false;
+    }
+
+    @ExportMessage(name = "isArrayElementReadable")
+    @ExportMessage(name = "isArrayElementModifiable")
+    boolean inBounds(long index) {
+        return 0 <= index && index < getSize();
     }
 
     @ExportMessage
-    boolean hasMetaObject() {
-        return true;
+    long getArraySize() {
+        return getSize();
     }
 
     @ExportMessage
-    Object getMetaObject() {
-        if (getSize() == 1) {
-            return Char1Type.SINGLETON;
+    public Object readArrayElement(long index) throws InvalidArrayIndexException {
+        try {
+            return getCharAt((int) index);
+        } catch (ArrayIndexOutOfBoundsException e) {
+            CompilerDirectives.transferToInterpreter();
+            throw InvalidArrayIndexException.create(index);
         }
-        return new CharType(getSize());
     }
 
     @ExportMessage
-    public String asString() {
-        return data;
+    public void writeArrayElement(long index, Object value) throws InvalidArrayIndexException {
+        try {
+            if (value instanceof Character) {
+                Character character = (Character)value;
+                setCharAt((int) index, character.charValue());                
+            } else {
+                throw new NotImplementedException();
+            }
+        } catch (ArrayIndexOutOfBoundsException e) {
+            CompilerDirectives.transferToInterpreter();
+            throw InvalidArrayIndexException.create(index);
+        }
     }
 
-    @ExportMessage
-    boolean isString() {
-        return true;
-    }
 }